adding classes for tabs
